controller/requirements.txt
pyautogui-0.9.54